SCP-3018: Restructured modules
